finalizing plugins, readme update
