+++ collectors/s3/getBucketAcl.js
@@ -1,6 +1,5 @@
 var index = require(__dirname + '/index.js');
 
 module.exports = function(AWSConfig, collection, callback) {
-	console.log('Begin getBucketAcl');
 	index('getBucketAcl', false, AWSConfig, collection, callback);
 };
