+++ exports.js
@@ -16,7 +16,6 @@ module.exports = {
     'defaultSecurityGroup'          : require(__dirname + '/plugins/ec2/defaultSecurityGroup.js'),
     'elasticIpLimit'                : require(__dirname + '/plugins/ec2/elasticIpLimit.js'),
     'excessiveSecurityGroups'       : require(__dirname + '/plugins/ec2/excessiveSecurityGroups.js'),
-    // 'insecureCiphers'               : require(__dirname + '/plugins/ec2/insecureCiphers.js'),
     'instanceLimit'                 : require(__dirname + '/plugins/ec2/instanceLimit.js'),
     'openCIFS'                      : require(__dirname + '/plugins/ec2/openCIFS.js'),
     'openDNS'                       : require(__dirname + '/plugins/ec2/openDNS.js'),
@@ -34,7 +33,11 @@ module.exports = {
     'openVNCClient'                 : require(__dirname + '/plugins/ec2/openVNCClient.js'),
     'openVNCServer'                 : require(__dirname + '/plugins/ec2/openVNCServer.js'),
     'vpcElasticIpLimit'             : require(__dirname + '/plugins/ec2/vpcElasticIpLimit.js'),
-    
+    'classicInstances'              : require(__dirname + '/plugins/ec2/classicInstances.js'),
+    'flowLogsEnabled'               : require(__dirname + '/plugins/ec2/flowLogsEnabled.js'),
+    'vpcMultipleSubnets'            : require(__dirname + '/plugins/ec2/multipleSubnets.js'),
+
+    'insecureCiphers'               : require(__dirname + '/plugins/elb/insecureCiphers.js'),
 
     'accessKeysExtra'               : require(__dirname + '/plugins/iam/accessKeysExtra.js'),
     'accessKeysLastUsed'            : require(__dirname + '/plugins/iam/accessKeysLastUsed.js'),
@@ -68,10 +71,6 @@ module.exports = {
     'domainTransferLock'            : require(__dirname + '/plugins/route53/domainTransferLock.js'),
 
     'bucketAllUsersPolicy'          : require(__dirname + '/plugins/s3/bucketAllUsersPolicy.js'),
-
-    'classicInstances'              : require(__dirname + '/plugins/vpc/classicInstances.js'),
-    'flowLogsEnabled'               : require(__dirname + '/plugins/vpc/flowLogsEnabled.js'),
-    'vpcMultipleSubnets'            : require(__dirname + '/plugins/vpc/multipleSubnets.js'),
 
     // OPTIONAL Plugins
     // These plugins require additional IAM permissions beyond the SecurityAudit managed policy
