+++ plugins/cloudtrail/cloudtrailFileValidation.js
@@ -14,11 +14,11 @@ module.exports = {
 		var results = [];
 		var source = {};
 
-		async.each(helpers.regions.cloudtrail, function(region, cb){
+		async.each(helpers.regions.cloudtrail, function(region, rcb){
 			var describeTrails = helpers.addSource(cache, source,
 				['cloudtrail', 'describeTrails', region]);
 
-			if (!describeTrails) return cb();
+			if (!describeTrails) return rcb();
 
 			if (describeTrails.err || !describeTrails.data) {
 				helpers.addResult(results, 3, 'Unable to query for CloudTrail file validation status', region);
@@ -40,7 +40,7 @@ module.exports = {
 			} else {
 				helpers.addResult(results, 2, 'CloudTrail is enabled but is not properly configured', region);
 			}
-			cb();
+			rcb();
 		}, function(){
 			callback(null, results, source);
 		});
