+++ plugins/ec2/classicInstances.js
